Adds Global  setup fixture
